requirements.txt
aiogram
aiogram-media-group
python-dotenv